requirements.txt
aniso8601==3.0.2
atomicwrites==1.1.5
attrs==18.1.0
Django==2.0.7
django-cors-headers==2.3.0
django-grappelli==2.11.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.8.16
graphene==2.1.2
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
more-itertools==4.2.0
pluggy==0.6.0
promise==2.1
psycopg2==2.7.5
py==1.5.4
PyHamcrest==1.9.0
pytest==3.6.3
python-dateutil==2.7.3
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
text-unidecode==1.2
typing==3.6.4
